% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TablaGenAnualCompleta.R
\name{TablaGenAnualCompleta}
\alias{TablaGenAnualCompleta}
\title{aaaaa}
\usage{
TablaGenAnualCompleta(
  Generacion,
  AñoBase = 2012,
  Tabla,
  Genero = FALSE,
  interes
)
}
\description{
aaaaa
}
